actual page commit
+ actual page code : 80%
+ actual page text english version : 80%
- removed useless fonts
- moved some css around (but you hadn't seen it)

been struggling with sass, time and what not.
The html code looks quite nice to my eyes and now I remember why nobody
crafts pages directly in code anymore.
Best exercise so far.

There is a ton of styling to do but the tedious part of writing abour
myself is over